Extract defaultReporter and selectReporter into separate module
